refactor: move fetch wrapper to lib folder and remove get pitch, remove utils folder
