--- client/src/utils/get-pitch.ts
@@ -1,5 +1,0 @@
-import type { PianoKey } from '~/store/piano-keys/slice';
-
-export const getPitch = (pianoKey: PianoKey) => {
-  return `${pianoKey.note}${pianoKey.octave}`;
-};
